week-7/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
pg8000==1.30.1
python-dotenv==1.0.0
requests==2.32.3